Added support for URLS to be specified as environment variables, kept official values as the defaults.
Added an example .env file, which can be used on custom apps and loaded as env vars, to override servers configuration.

Made email services and sms services to have objects as parameters, and not a lot of parameters on their signatures. This helps us avoid doing large chagnes if another parameter is added.

Added the new data, custom account id, and custom data.